requirements.txt
# Cursor Rules Optimization Dependencies
# Version: 1.0.0

# Core dependencies
pyyaml>=6.0.1      # YAML parsing for metadata
psutil>=5.9.0      # System monitoring for performance metrics
click>=8.1.0       # CLI framework for scripts
tiktoken>=0.5.0    # Token counting for LLM context
networkx>=3.0      # Graph operations for dependency analysis
matplotlib>=3.7.0  # Visualization for dependency graphs

# Development dependencies
pytest>=7.4.0      # Testing framework
black>=23.12.0     # Code formatting
ruff>=0.1.0        # Linting
mypy>=1.8.0        # Type checking

# Documentation
markdown>=3.5.0    # Markdown processing
mermaid-cli>=0.1.0 # Diagram generation

# Optional performance monitoring
prometheus-client>=0.19.0  # Metrics export
